+++ src/Anonymous/AnonymousSettingFactory.php
@@ -16,12 +16,12 @@ class AnonymousSettingFactory
         static::$typeMap[$alias] = $callback;
     }
 
-    public static function anonymous(string $type, string $key, mixed $defaultValue, Field $fieldOptions, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null): AnonymousSetting
+    public static function anonymous(string $type, string $key, mixed $defaultValue, ?Field $fieldOptions = null, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null): AnonymousSetting
     {
         return static::create($type, $key, $defaultValue, $fieldOptions, $groups, $rules, $resolveIdUsing);
     }
 
-    public static function create(string $type, string $key, mixed $defaultValue, Field $fieldOptions, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null): AnonymousSetting
+    public static function create(string $type, string $key, mixed $defaultValue, ?Field $fieldOptions = null, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null): AnonymousSetting
     {
         $setting = static::make($type, $key, $defaultValue, $fieldOptions, $groups, $rules, $resolveIdUsing);
 
@@ -30,7 +30,7 @@ class AnonymousSettingFactory
         return $setting;
     }
 
-    public static function make(string $type, string $key, mixed $defaultValue, Field $fieldOptions, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null): AnonymousSetting
+    public static function make(string $type, string $key, mixed $defaultValue, ?Field $fieldOptions = null, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null): AnonymousSetting
     {
         if($resolveIdUsing === null) {
             $resolveIdUsing = array_key_exists($type, static::$typeMap) ? static::$typeMap[$type] : throw new \Exception(sprintf('Anonymous setting type [%s] was not found.', $type));