+++ src/Decorators/BaseSettingServiceDecorator.php
@@ -92,17 +92,17 @@ class BaseSettingServiceDecorator implements SettingService
         return $this->baseService->getSettingByKey($key);
     }
 
-    public function create(string $type, string $key, mixed $defaultValue, Field $fieldOptions, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null): Setting
+    public function create(string $type, string $key, mixed $defaultValue, ?Field $fieldOptions = null, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null): Setting
     {
         return $this->baseService->create($type,  $key, $defaultValue, $fieldOptions, $groups, $rules, $resolveIdUsing);
     }
 
-    public function createUser(string $key, mixed $defaultValue, Field $fieldOptions, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null): Setting
+    public function createUser(string $key, mixed $defaultValue, ?Field $fieldOptions = null, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null): Setting
     {
         return $this->baseService->createUser($key, $defaultValue, $fieldOptions, $groups, $rules, $resolveIdUsing);
     }
 
-    public function createGlobal(string $key, mixed $defaultValue, Field $fieldOptions, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null): Setting
+    public function createGlobal(string $key, mixed $defaultValue, ?Field $fieldOptions = null, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null): Setting
     {
         return $this->baseService->createGlobal($key, $defaultValue, $fieldOptions, $groups, $rules, $resolveIdUsing);
     }