+++ src/Setting.php
@@ -24,9 +24,9 @@ use Settings\Store\Query;
  * @method static Query withGlobalType() Only global settings
  * @method static Query withUserType() Only user settings
  * @method static void alias(string $alias, string $key) Alias the $key with $alias
- * @method static \Settings\Contracts\Setting create(string $type, string $key, mixed $defaultValue, Field $fieldOptions, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null) Create and register a new setting
- * @method static \Settings\Contracts\Setting createUser(string $key, mixed $defaultValue, Field $fieldOptions, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null) Create and register a user setting
- * @method static \Settings\Contracts\Setting createGlobal(string $key, mixed $defaultValue, Field $fieldOptions, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null) Create and register a global setting
+ * @method static \Settings\Contracts\Setting create(string $type, string $key, mixed $defaultValue, ?Field $fieldOptions = null, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null) Create and register a new setting
+ * @method static \Settings\Contracts\Setting createUser(string $key, mixed $defaultValue, ?Field $fieldOptions = null, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null) Create and register a user setting
+ * @method static \Settings\Contracts\Setting createGlobal(string $key, mixed $defaultValue, ?Field $fieldOptions = null, array $groups = ['default'], array|string $rules = [], ?\Closure $resolveIdUsing = null) Create and register a global setting
  * @method static void loadSetting(string $key) Share a setting value with the frontend
  * @method static void loadManySettings(array $keys) Share many setting values with the frontend
  *