+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.4] - 2022-02-19
+
+### Changed
+- Made forms optional
+- Let you put `null` for cache ttl to cache forever
+
 ## [v1.0.3] - 2022-02-15
 
 ### Added